# Conflicts:
#	src/app.py

+++ src/dataAcces/transfer.py
@@ -76,6 +76,7 @@ class TransferDataAccess:
         """
         cursor = self.dbconnect.get_cursor()
         if transfer:  # If it's a transfer, we take the middle of the 2 sids
+            print(oid)
             cursor.execute('SELECT idTo,toType,idFrom, fromtype FROM transfer WHERE id=%s;', (oid,))
             ids = cursor.fetchone()
             sidTo = self.translatePosition(ids[0], ids[1])  # This can go recursively for transfers on transfers
@@ -187,7 +188,7 @@ class TransferDataAccess:
             return None
         return Transfer(tid, data[1], data[2], data[3], data[4], data[5], data[6], data[7])
 
-    def returnToBase(self, transfer: Transfer, timer_data_access, soldier_data_acces, package_data_acces):
+    def returnToBase(self, transfer: Transfer, timer_data_access, soldier_data_acces, package_data_acces, timer=None):
         """
         Helper function to delete the current transfer and make a new transfer towards home (idFrom)
         :param package_data_acces:
@@ -200,7 +201,9 @@ class TransferDataAccess:
         cursor.execute("SELECT id FROM timer WHERE oid=%s and type IN('attack','outpost','transfer')", (oldTid,))
         originalTimerID = cursor.fetchone()  # Already get the original timer ID before we make any changes
 
-        if originalTimerID is None:  # The transfer was an espionage; will be handled separately
+        if originalTimerID is None and timer is not None:
+            pass
+        elif originalTimerID is None:  # Spionage is handled differently
             return
         else:
             originalTimerID = originalTimerID[0]
@@ -208,16 +211,13 @@ class TransferDataAccess:
         if transfer.toType:  # If we went to a transfer x, the start location will be the end of x
             cursor.execute('SELECT idTo FROM transfer WHERE id=%s;', (transfer.idTo,))
             idTo = cursor.fetchone()[0]
-            # TODO Unsure if this works in frontend, since it goes towards the same settlement now
         else:  # We went to a settlement, now leaving from there: transfer data doesn't need to be adjusted
             idTo = transfer.idTo
         # Make a resource transfer back home
         cursor.execute(
-            'INSERT INTO transfer(idto, totype, idfrom, fromtype, discovered, pid, pname) VALUES (%s,%s,%s,%s,%s,%s,%s)',
+            'UPDATE transfer SET idto = %s ,  totype = %s , idfrom = %s , fromtype = %s , discovered = %s , pid = %s , pname = %s WHERE id = %s',
             (transfer.idFrom, False, idTo, False, True, transfer.pid,
-             transfer.pname))  # Insert new transfer into the database
-        cursor.execute('SELECT max(id) FROM transfer;')  # Retrieve the tid
-        transfer.id = cursor.fetchone()[0]
+             transfer.pname, transfer.id))  # Insert new transfer into the database
 
         # Add a new timer
         start, stop, duration = self.calculateDuration(soldier_data_acces.getTroops(transfer.pid, 'package'),
@@ -227,19 +227,19 @@ class TransferDataAccess:
         timer = Timer(None, transfer.id, 'transfer', start, stop, duration, transfer.idTo)
         timer_data_access.insertTimer(timer)
 
-        if transfer.toType:
-            # We need to call this recursively on any transfers linked to this one
-            cursor.execute('SELECT id FROM transfer WHERE totype=True and idTo=%s;',
-                           (transfer.id,))
-            transfers = cursor.fetchall()
-            for tid in transfers:  # Send them back to where they came from
-                self.returnToBase(self.instantiateTransfer(tid[0]), timer_data_access,
-                                  soldier_data_acces, package_data_acces)
-                self.dbconnect.commit()
+        # if transfer.toType:
+        #     # We need to call this recursively on any transfers linked to this one
+        #     cursor.execute('SELECT id FROM transfer WHERE totype=True and idTo=%s;',
+        #                    (transfer.id,))
+        #     transfers = cursor.fetchall()
+        #     for tid in transfers:  # Send them back to where they came from
+        #         self.returnToBase(self.instantiateTransfer(tid[0]), timer_data_access,
+        #                           soldier_data_acces, package_data_acces)
+        #         self.dbconnect.commit()
 
-        # Delete the old transfer and old timer in the database (package is recycled)
-        cursor.execute('DELETE FROM transfer WHERE id=%s;', (oldTid,))
-        cursor.execute('DELETE FROM timer WHERE id=%s;', (originalTimerID,))
+        # Delete the old timer in the database (package and transfer is recycled)
+        if originalTimerID is not None:
+            cursor.execute('DELETE FROM timer WHERE id=%s;', (originalTimerID,))
 
         # Commit Data
         self.dbconnect.commit()
